+++ netwerk/cookie/CookieServiceChild.cpp
@@ -136,6 +136,7 @@ void CookieServiceChild::TrackCookieLoad(nsIChannel* aChannel) {
 
   bool isForeign = false;
   bool isTrackingResource = false;
+  bool isSocialTrackingResource = false;
   bool firstPartyStorageAccessGranted = false;
   uint32_t rejectedReason = 0;
   nsCOMPtr<nsIURI> uri;
@@ -148,6 +149,7 @@ void CookieServiceChild::TrackCookieLoad(nsIChannel* aChannel) {
       do_QueryInterface(aChannel);
   if (classifiedChannel) {
     isTrackingResource = classifiedChannel->IsTrackingResource();
+    isSocialTrackingResource = classifiedChannel->IsSocialTrackingResource();
     // Check first-party storage access even for non-tracking resources, since
     // we will need the result when computing the access rights for the reject
     // foreign cookie behavior mode.
@@ -172,6 +174,7 @@ void CookieServiceChild::TrackCookieLoad(nsIChannel* aChannel) {
   bool isSafeTopLevelNav = NS_IsSafeTopLevelNav(aChannel);
   bool isSameSiteForeign = NS_IsSameSiteForeign(aChannel, uri);
   SendPrepareCookieList(uriParams, isForeign, isTrackingResource,
+                        isSocialTrackingResource,
                         firstPartyStorageAccessGranted, rejectedReason,
                         isSafeTopLevelNav, isSameSiteForeign, attrs);
 }
@@ -265,9 +268,9 @@ void CookieServiceChild::PrefChanged(nsIPrefBranch* aPrefBranch) {
 
 void CookieServiceChild::GetCookieStringFromCookieHashTable(
     nsIURI* aHostURI, bool aIsForeign, bool aIsTrackingResource,
-    bool aFirstPartyStorageAccessGranted, uint32_t aRejectedReason,
-    bool aIsSafeTopLevelNav, bool aIsSameSiteForeign, nsIChannel* aChannel,
-    nsACString& aCookieString) {
+    bool aIsSocialTrackingResource, bool aFirstPartyStorageAccessGranted,
+    uint32_t aRejectedReason, bool aIsSafeTopLevelNav, bool aIsSameSiteForeign,
+    nsIChannel* aChannel, nsACString& aCookieString) {
   nsCOMPtr<nsIEffectiveTLDService> TLDService =
       do_GetService(NS_EFFECTIVETLDSERVICE_CONTRACTID);
   NS_ASSERTION(TLDService, "Can't get TLDService");
@@ -304,7 +307,7 @@ void CookieServiceChild::GetCookieStringFromCookieHashTable(
 
   CookieStatus cookieStatus = nsCookieService::CheckPrefs(
       cookieSettings, aHostURI, aIsForeign, aIsTrackingResource,
-      aFirstPartyStorageAccessGranted, VoidCString(),
+      aIsSocialTrackingResource, aFirstPartyStorageAccessGranted, VoidCString(),
       CountCookiesFromHashTable(baseDomain, attrs), attrs, &aRejectedReason);
 
   if (cookieStatus != STATUS_ACCEPTED &&
@@ -471,12 +474,14 @@ nsresult CookieServiceChild::GetCookieStringInternal(
   }
 
   bool isTrackingResource = false;
+  bool isSocialTrackingResource = false;
   bool firstPartyStorageAccessGranted = false;
   uint32_t rejectedReason = 0;
   nsCOMPtr<nsIClassifiedChannel> classifiedChannel =
       do_QueryInterface(aChannel);
   if (classifiedChannel) {
     isTrackingResource = classifiedChannel->IsTrackingResource();
+    isSocialTrackingResource = classifiedChannel->IsSocialTrackingResource();
     // Check first-party storage access even for non-tracking resources, since
     // we will need the result when computing the access rights for the reject
     // foreign cookie behavior mode.
@@ -490,9 +495,9 @@ nsresult CookieServiceChild::GetCookieStringInternal(
   bool isSameSiteForeign = NS_IsSameSiteForeign(aChannel, aHostURI);
 
   GetCookieStringFromCookieHashTable(
-      aHostURI, isForeign, isTrackingResource, firstPartyStorageAccessGranted,
-      rejectedReason, isSafeTopLevelNav, isSameSiteForeign, aChannel,
-      aCookieString);
+      aHostURI, isForeign, isTrackingResource, isSocialTrackingResource,
+      firstPartyStorageAccessGranted, rejectedReason, isSafeTopLevelNav,
+      isSameSiteForeign, aChannel, aCookieString);
 
   return NS_OK;
 }
@@ -516,12 +521,14 @@ nsresult CookieServiceChild::SetCookieStringInternal(
   }
 
   bool isTrackingResource = false;
+  bool isSocialTrackingResource = false;
   bool firstPartyStorageAccessGranted = false;
   uint32_t rejectedReason = 0;
   nsCOMPtr<nsIClassifiedChannel> classifiedChannel =
       do_QueryInterface(aChannel);
   if (classifiedChannel) {
     isTrackingResource = classifiedChannel->IsTrackingResource();
+    isSocialTrackingResource = classifiedChannel->IsSocialTrackingResource();
     // Check first-party storage access even for non-tracking resources, since
     // we will need the result when computing the access rights for the reject
     // foreign cookie behavior mode.
@@ -556,7 +563,7 @@ nsresult CookieServiceChild::SetCookieStringInternal(
   // Asynchronously call the parent.
   if (CanSend()) {
     SendSetCookieString(hostURIParams, channelURIParams, optionalLoadInfoArgs,
-                        isForeign, isTrackingResource,
+                        isForeign, isTrackingResource, isSocialTrackingResource,
                         firstPartyStorageAccessGranted, rejectedReason, attrs,
                         cookieString, nsCString(aServerTime), aFromHttp);
   }
@@ -571,7 +578,7 @@ nsresult CookieServiceChild::SetCookieStringInternal(
 
   CookieStatus cookieStatus = nsCookieService::CheckPrefs(
       cookieSettings, aHostURI, isForeign, isTrackingResource,
-      firstPartyStorageAccessGranted, aCookieString,
+      isSocialTrackingResource, firstPartyStorageAccessGranted, aCookieString,
       CountCookiesFromHashTable(baseDomain, attrs), attrs, &rejectedReason);
 
   if (cookieStatus != STATUS_ACCEPTED &&